requirements.txt
websockets-server==0.4
websocket-client
zmq
flask
flask-cors